Generacion de token y proteccion de endpoints

+++ src/main/java/com/procesos/negocio/andres/controllers/UsuarioController.java
@@ -1,6 +1,7 @@
 package com.procesos.negocio.andres.controllers;
 import com.procesos.negocio.andres.models.Usuario;
 import com.procesos.negocio.andres.services.UsuarioService;
+import com.procesos.negocio.andres.utils.JWTUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,47 +17,71 @@ import java.util.Map;
 public class UsuarioController {
     @Autowired
     private UsuarioService usuarioService;
-
+    @Autowired
+    private JWTUtil jwtUtil;
 
     @GetMapping(value = "/usuario/{id}")
-    public ResponseEntity getUsuario(@PathVariable Long id) {
-        return usuarioService.getUserById(id);
+    public ResponseEntity getUsuario(@PathVariable Long id, @RequestHeader(value= "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+        }
 
+        return usuarioService.getUserById(id);
     }
 
     @PostMapping("/usuario")
     public ResponseEntity crearUsuario(@Valid @RequestBody  Usuario usuario) {
+
         return usuarioService.createUser(usuario);
 
     }
 
     @GetMapping("/usuarios")
-    public ResponseEntity listarUsuario() {
+    public ResponseEntity listarUsuario(@RequestHeader(value= "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+        }
         return usuarioService.allUsers();
     }
 
     @GetMapping("/usuario/{nombre}/{apellidos}")
-    public ResponseEntity listaPorNombreApellidos(@PathVariable String nombre, @PathVariable String apellidos) {
+    public ResponseEntity listaPorNombreApellidos(@PathVariable String nombre, @PathVariable String apellidos,@RequestHeader(value= "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+        }
         return usuarioService.allUsersByNameAndLastName(nombre,apellidos);
     }
 
     @GetMapping("/usuario/nombre/{nombre}")
-    public ResponseEntity listaPorNombre(@PathVariable String nombre) {
+    public ResponseEntity listaPorNombre(@PathVariable String nombre,@RequestHeader(value= "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+        }
         return usuarioService.allUsersByName(nombre);
     }
 
     @GetMapping("/usuario/apellidos/{apellidos}")
-    public ResponseEntity listaPorApellidos(@PathVariable String apellidos) {
+    public ResponseEntity listaPorApellidos(@PathVariable String apellidos,@RequestHeader(value= "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+        }
         return usuarioService.allUsersByLastName(apellidos);
     }
 
     @PutMapping("/usuario/{id}")
-    public ResponseEntity editarUsuario(@PathVariable Long id, @Valid @RequestBody  Usuario usuario) {
+    public ResponseEntity editarUsuario(@PathVariable Long id, @Valid @RequestBody  Usuario usuario,@RequestHeader(value= "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+        }
         return usuarioService.editUser(id, usuario);
     }
 
     @DeleteMapping("/usuario/{id}")
-    public ResponseEntity eliminarUsuario(@PathVariable Long id) {
+    public ResponseEntity eliminarUsuario(@PathVariable Long id,@RequestHeader(value= "Authorization") String token) {
+
+        if(jwtUtil.getKey(token)==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+        }
         return usuarioService.deleteUserById(id);
     }
     @ResponseStatus(HttpStatus.BAD_REQUEST)
